+++ text_transform.py
@@ -1,15 +1,15 @@
-import re
 from pathlib import Path
 from typing import List
 
 from get_files_of_name_tags import get_filenames_of_name_tags
+from replace_relative_tags_with_absolute_tags_in_file import (
+    replace_relative_tags_with_absolute_tags_in_file,
+)
 
 from src.recipes import split_file_with_separators
 from src.search_configuration import SearchConfiguration
 from src.search_configurations_reader import SearchConfigurationsReader
-from src.text_file import TextFile
 from src.text_file_io import TextFileIO
-from src.text_processor import TextProcessor
 
 INPUT_TEXT_FILE_PATH = Path(".sandbox/CppCoreGuidelines.md")
 INPUT_SEPARATORS_FILE_PATH = Path(".sandbox/separators.json")
@@ -36,31 +36,12 @@ absolute_tag_of_relative_tag = {
 }
 
 
-def replace_relative_tag_by_absolute_tag_in_match(
-    match_object: re.Match,
-) -> str:
-    relative_name = match_object.groupdict()["relative"]
-    absolute_name = absolute_tag_of_relative_tag[relative_name.lstrip("#")]
-    return match_object.string.replace(relative_name, absolute_name)
-
-
-input_file_text_processor = TextProcessor(input_file.text)
-
-for relative_tag, absolute_tag in absolute_tag_of_relative_tag.items():
-    pattern = SearchConfiguration(
-        name="name_tag",
-        regex_pattern=rf".*\[.*\]\((?P<relative>#{relative_tag})\).*",
-    )
-    input_file_text_processor = TextProcessor(
-        input_file_text_processor.substitute_pattern_with_replacement(
-            pattern=pattern,
-            replacement=replace_relative_tag_by_absolute_tag_in_match,
-        )
-    )
+edited_file = replace_relative_tags_with_absolute_tags_in_file(
+    input_file, absolute_tag_of_relative_tag
+)
 
 separated_files_with_absolute_references = split_file_with_separators(
-    TextFile(text=input_file_text_processor.get_text(), path=Path("")),
-    file_separators,
+    edited_file, file_separators
 )
 
 for file_a in separated_files_with_absolute_references:
